Move settings management commands to subapp
