+++ tests/janitor/management/commands/test_settings_shake.py
@@ -4,7 +4,7 @@ import pytest
 
 from django.core.management import call_command
 
-from directory_components.management.commands import helpers, settings_shake
+from directory_components.janitor.management.commands import helpers, settings_shake
 
 
 @pytest.fixture(autouse=True)
